+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Pyxel doesn't use SemVer anymore, since most minor releases have at least minor 
 This release brings a number of bugfixes, documentation and improvements.
 
 ### Core
+* Removed deprecated `Detector.set_output_dir` method and `Detector.output_dir` property.
+  (See [!759](https://gitlab.com/esa/pyxel/-/merge_requests/749)).
 
 ### Documentation